# Conflicts:
#	src/main/java/com/esprit/examen/entities/Formateur.java
#	src/main/java/com/esprit/examen/repositories/FormateurRepository.java
#	src/main/java/com/esprit/examen/services/SessionService.java
#	src/test/java/com/esprit/examen/SessionTest.java

+++ src/main/java/com/esprit/examen/entities/Session.java
@@ -28,7 +28,7 @@ public class Session implements Serializable{
     Formateur formateur;
 	@ManyToMany(fetch = FetchType.EAGER)
 	Set<Cours> cours;
-	private Long price;
+	private Double price;
 
 	private Long salaireF;
 
@@ -99,7 +99,7 @@ public class Session implements Serializable{
 
 
 
-	public Session(Date dateDebut, Date dateFin, Long duree, String description, Formateur formateur, Set<Cours> cours, Long price) {
+	public Session(Date dateDebut, Date dateFin, Long duree, String description, Formateur formateur, Set<Cours> cours, Double price) {
 		this.dateDebut = dateDebut;
 		this.dateFin = dateFin;
 		this.duree = duree;
@@ -108,7 +108,7 @@ public class Session implements Serializable{
 		this.cours = cours;
 		this.price = price;
 	}
-	public Session(Date dateDebut, Date dateFin, Long duree, String description, Set<Cours> cours, Long price) {
+	public Session(Date dateDebut, Date dateFin, Long duree, String description, Set<Cours> cours, Double price) {
 		this.dateDebut = dateDebut;
 		this.dateFin = dateFin;
 		this.duree = duree;
